notifications-api/requirements.txt
aiofiles==22.1.0
alembic==1.9.1
anyio==3.6.2
asyncio==3.4.3
click==8.1.3
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
pydantic==1.10.4
pydash==5.1.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a31
sqlmodel==0.0.8
sse-starlette==1.2.1
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
pyjwt[crypto]==2.6.0
